dt-bindings: mailbox: omap: Update binding for DRA82x SoCs
Update the existing OMAP Mailbox binding to also include the
info for DRA82x SoCs. The same compatible from AM65x SoCs is
reused for DRA82x SoCs as well. For now, the interrupts are
expected to be a direct GIC_SPI interrupt until the TI-SCI
irqchip driver is added.

Signed-off-by: Suman Anna <s-anna@ti.com>