TST: arm64: dts: ti: dra822: Add mailbox loopback devices for testing
Add some loopback sub-mailbox devices within each of the Mailbox Clusters
0 through 4 on the DRA82x SoCs for testing. The loopback device uses the
same mailbox fifo for sending and receiving messages on the MPU itself,
facilitating a unit-test. The last FIFO within each cluster is chosen for
the loopback to not conflict with the actual FIFOs at the beginning that
will be used to communicate with the remote processors.

Signed-off-by: Suman Anna <s-anna@ti.com>