ti_config_fragments: *android.cfg: Enable Firmware fallback mechanism
Enable firmware fallback mechanism to allow to overcome failures to do
a direct filesystem (rootfs)  when userspace is not ready.

In Android, we have seen scenarios of firmwares not available, because vendor
partition was not ready due to the early mounting requirements.

The fallback mechanism has been particularly sucessful in loading
the firmware from android <rootfs>/vendor/firmware/*

Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>