firmware: ti_sci: Rework the irq_ops to configure inta and intr separately
Currently the irq route form IP -> INTA -> INTR -> GIC is configured in
a single message and happens in inta driver. This will not not work when
ABI 3.0 is introduced. So, update irq_ops such that INTA configuration
happens in one message and INTR configuration happens in a different
message for the above mentioned irq route.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>