UPSTREAM: arm64: dts: ti: k3-am62: Enable EPWM nodes at the board level
EPWM nodes defined in the top-level AM62x SoC dtsi files are incomplete
and will not be functional unless they are extended with pinmux
information.

As the pinmux is only known at the board integration level, these
nodes should only be enabled when provided with this information.

Disable the EPWM nodes in the dtsi files and only enable the ones that
are actually pinned out on a given board.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Bryan Brattlof <bb@ti.com>
Link: https://lore.kernel.org/r/20221018211533.21335-5-afd@ti.com
(cherry picked from commit 6e5474826606a783dd969467f93b0c6a56c39780)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>