UPSTREAM: arm64: dts: ti: k3-am62: MDIO pinmux should belong to the MDIO node
Although usually integrated as a child of an Ethernet controller, MDIO
IP has an independent pinout. This pinout should be controlled by
the MDIO node (so if it was to be disabled for instance, the pinmux
state would reflect that).

Move the MDIO pins pinmux to the MIDO nodes.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Bryan Brattlof <bb@ti.com>
Link: https://lore.kernel.org/r/20221018211533.21335-7-afd@ti.com
(cherry picked from commit 8709e0442c79c03f7067d6e7464d669504f76527)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>