UPSTREAM: arm64: dts: ti: k3-am62: Enable MDIO nodes at the board level
MDIO nodes defined in the top-level AM62x SoC dtsi files are incomplete
and will not be functional unless they are extended with a pinmux.

As the attached PHY is only known about at the board integration level,
these nodes should only be enabled when provided with this information.

Disable the MDIO nodes in the dtsi files and only enable the ones that
are actually pinned out on a given board.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Bryan Brattlof <bb@ti.com>
Link: https://lore.kernel.org/r/20221018211533.21335-8-afd@ti.com
(cherry picked from commit 1f737012ce7bd92aecf7d6b1e17742dc7cae7bee)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>