UPSTREAM: arm64: dts: ti: k3-j721e: Enable MCASP nodes at the board level
MCASP nodes defined in the top-level J721e SoC dtsi files are incomplete
and may not be functional unless they are extended with pinmux
information.

As the pinmux is only known at the board integration level, these
nodes should only be enabled when provided with this information.

Disable the MCASP nodes in the dtsi files and only enable the ones that
are actually pinned out on a given board.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Bryan Brattlof <bb@ti.com>
Link: https://lore.kernel.org/r/20221020160305.18711-4-afd@ti.com
(cherry picked from commit 256596ad158c80c8e4104a6310a411ec8829f9eb)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>