UPSTREAM: arm64: dts: ti: k3-j721s2: Enable I2C nodes at the board level
I2C nodes defined in the top-level J721s2 SoC dtsi files are incomplete
and will not be functional unless they are extended with pinmux
information.

As the pinmux is only known at the board integration level, these
nodes should only be enabled when provided with this information.

Disable the I2C nodes in the dtsi files and only enable the ones that
are actually pinned out on a given board.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Bryan Brattlof <bb@ti.com>
Link: https://lore.kernel.org/r/20221020160305.18711-13-afd@ti.com
(cherry picked from commit 0aef5131380628e2a9bfcbbcfed5d6477c8f9651)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>