TI: media: cadence: csi2rx: Add get_fmt and set_fmt pad ops
The format is needed to calculate the link speed for the external DPHY
configuration. It is not right to query the format from the source
subdev. Add get_fmt and set_fmt pad operations so that the format can be
configured and correct bpp be selected.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
(cherry picked from commit a332e50072c0c7ae8786c5603a22674a521b3a8b)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>