arm64: dts: ti: k3-am62a-main: Add node for Display SubSystem
Add DT node for the Display SubSystem on the am62a soc in cbass_main.
The DSS IP on this soc is compatible with the one on the am625 soc.

The DSS supports one each of video pipeline (vid) and video-lite
pipeline (vidl1). The OLDI TXes at the end of video port 1 (vp1) are
tied OFF in AM62A SoC, but the pipeline remains active. The video port 2
(vp2) outputs the DPI signals. Both the video ports are connected to the
pipelines via 2 identical overlay managers (ovr1 and ovr2).

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Reviewed-by: Devarsh Thakkar <devarsht@ti.com>