TI: net: ethernet: ti: am65-cpsw-qos: Add IET Frame preemption feature
This patch adds support for Intersperse Express Traffic (IET, defined
in P802.3br/D2.0 spec which later is included in IEEE 802.3 2018) Frame
preemption (FPE) feature available in CPSW h/w. Also add MAC Verify
to allow MAC layer to Verify if the peer supports IET MAC merge
layer or not. MAC merge layer is responsible for preempting the
transmission of frame from a preemptible queue if there is frame
waiting for transmission at a higher priority Express queue. The h/w
sends an initial segment of the frame satisfying min fragment size
requirement and then schedule frame from the Express queue for
transmission. Finally when no more frames available at the Express queue,
it will resume transmission of remaining segments of the frame of
the preemptible queue which was preempted. At the peer end, the
segments are re-assembled and delivered to the MAC interface.

IET FPE is configured for a port through ethtool --set-priv-flags command.
Driver configures IET FPE for a port in ndo_open() if the
fpe_configured flag is set. It schedules a worker to do the MAC/Verify
process as soon as the Link is up and mac_verify_configured is set.
It resets the LINKFAIL bit and check if the Verify succeeds or not.
On failure, the MAC Verify state machine is reset and process repeats.
As the peer comes up at different instances, it attempts multiple time
until succeeds or times out. This patch moves am65_est_cmd_ns_to_cnt()
function to the top of the file to re-use the same in the IET code as
the Verify timeout configured in the IET register is in wired clocks.
Also initializes IET in am65_cpsw_nuss_ndo_slave_open() and cleanup the
same in am65_cpsw_nuss_ndo_slave_stop()

Driver configures lower queues (Q0-QN-2, where N is the maximum number
of queues configured) as preemptible queues by programming the
PN_REG_IET_CTRL register if the MAC Verify succeeds or if the force mode
is enabled (MAC Verify disabled through priv flag). Also configures the
TX_RX_MAX_BLKS for IET in am65_cpsw_nuss_ndo_slave_open() if IET FPE is
configured by user.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Roger Quadros <rogerq@kernel.org>
Acked-by: Ravi Gunasekaran <r-gunasekaran@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
(cherry picked from commit 40e08b2aa7988da80aabc2425e0171f80109ba71)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>