TI: HACK: arm64: dts: ti: k3-am68-sk-base-board: Set status okay to enable DP and DSI
Enable DSI display for AM68 SK by setting status for all required nodes
for DP0 as "okay". To connect the panel as a sink, move the panel node to
the DSI bridge controller as a child node under the aux-bus node.

Currently we are using panel instead of dp-connector as a HACK.
Once we have get_edid functionality working, we will switch to
dp-connector node.

Fixes: 84f5762fa656 ("arm64: dts: ti: k3-am68-sk: Add DSS, HDMI, and DSI to eDP bridge")
Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>
Reviewed-by: Jayesh Choudhary <j-choudhary@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
(cherry picked from commit f31ead78270337647241233e36818052edb33202)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>