ANDROID: TI: add custom lz4 compression for Image.lz4
The kernel build system is capable of generating a compressed Image.lz4.
That image, however, is compressed with the legacy (-l) option, which is
incompatible with with TI U-Boot v2021.01.

To be compatible with TI U-Boot v2021, we need the Image.lz4 to be
generated with:

  lz4 -f -12 --favor-decSpeed /arch/arm64/boot/Image /arch/arm64/boot/Image.lz4

Add a custom genrule for this.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>