TI: remoteproc: k3-m4: set as wakeup capable but keep disabled by default
commit 3c47db3916a0 ("remoteproc: k3-m4: Enable MCU wakeup support")
enabled k3-m4 rproc as a wakeup source, however let's keep it as wakeup
capable, but not wakeup enabled by default. The user is expected to
explicitly go and set this as enabled if they wish to use it as a wakeup
source.

Suggested-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Dhruva Gole <d-gole@ti.com>
(cherry picked from commit 6d4677519856f5c33ad9ff6c3471ef126b8804cf)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>