TI: arm64: dts: ti: k3-am62x-lp-sk: Disable PRU
Drop the PRU node from the devie tree on the AM62 LP SK boards, as
the AM62x SoC variants populated on those boards do not have one.

Attempting to access the PRU will result in a crash at kernel bootup.

At this time, we don't have any code that is able to identify the
absence of the PRU by reading the SoC registers. Do this in the board
DTS file for now, since all of the AMC package variants will not have
the PRU.

Signed-off-by: Anand Gadiyar <gadiyar@ti.com>
Cc: Andrew Davis <afd@ti.com>
Cc: Vignesh Raghavendra <vigneshr@ti.com>
Acked-by: Andrew Davis <afd@ti.com>
[hnagalla@ti.com: forward port from 5.10 to 6.1]
Signed-off-by: Hari Nagalla <hnagalla@ti.com>
(cherry picked from commit 15dabfa67bacf63936dbd79184d4d5740c385337)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>