TI: HACK: drm/bridge: ite-it66121: Add drm_connector support
Add support for DRM connector and make the driver support the older
format of attaching connectors onto the encoder->bridge->connector
chain. This way, the bridge driver is responsible for creating and
attaching DRM connectors instead of the encoder.

This is a hack patch as the new method has not been introduced for the
tidss driver in the ti-6.1 kernel.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
(cherry picked from commit 85e6b8d401de1286d5360f4fa3dfa4b88052fd8d)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>