TI: media: cadence: csi2rx: Fix check for running bit
The running bit (31) of STREAM_STATUS register should go to idle (0)
when stream stop operation is finished. That was the intention in the
original patch but we were checking for it to be running (1) due to a
typo.

Fixes: c4cae42493cb ("media: cadence: csi2rx: Set the STOP bit when stopping a stream")
Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
(cherry picked from commit e5707c4f7e88d483426ab59dc0da314711203a60)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>