TI: arm64: dts : ti: k3-am69-sk: Add DDR carveout memory nodes for R5F
commit 567f75ab67d0d1b5ef4a2ca9d28f23e0876e1d61 upstream.

Two carveout reserved memory nodes each have been added for each of the
R5F remote processor device within both the MCU and MAIN domains for the
TI K3 AM69 SK boards. These nodes are assigned to the respective rproc
device nodes as well. The first region will be used as the DMA pool for
the rproc device, and the second region will furnish the static carveout
regions for the firmware memory.

The current carveout addresses and sizes are defined statically for each
device. The R5F processors do not have an MMU, and as such require the
exact memory used by the firmwares to be set-aside. The firmware images
do not require any RSC_CARVEOUT entries in their resource tables either
to allocate the memory for firmware memory segments.

Note that the R5F1 carveouts are needed only if the R5F cluster is
running in Split (non-LockStep) mode. The reserved memory nodes can be
disabled later on if there is no use-case defined to use the corresponding
remote processor.

Change-Id: I26ba5bfbe5ebaebae7ece9fdfa3ea16502c4f953
Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>
Signed-off-by: Apurva Nandan <a-nandan@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>
Link: https://lore.kernel.org/r/20231001181417.743306-9-a-nandan@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
(cherry picked from commit 319d4d153300d9b80240e8237dd65524e1796184)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>