TI: media: platform: img: vxd: Address stream compliance and add new user controls for buf allocations
Fix the compliance problem with the driver, and incorrect constraint was
being passed down into the IMG firmware.  Artificially setting
max_dec_frame_buffering to 1.  This renders some streams unplayable if
they have need more display picture buffers for reordering or if they
have higher reference frame dependencies.  Setting this value to 0 will
let the firmware decide the number of DPB's it will use.

This change then forces an increase in default DPB memory allocations
which may not be entirely practical in an embedded system, so this patch
also provides some extra custom V4L2 controls so that the memory
requirements can be tailored for a particular use case.

The DPB algorithm is now:
  num_ref_frames + ((num_cores * slots_per_core) - 1) + display_pipeline_length

num_cores is 2, and slots_per_core is 2 on J721e and
display_pipeline_length is 3.

num_core and slots_per core can be reduced, but IMG says this will
impact performance because the pipelines on the video hardware may not
be used as efficiently.   display_pipeline_length will depend on a
number of factors and this value can be tinkered with depending on the
usecase.

Add these controls into the driver

        max_dec_frame_buffering 0x00981901 (int)    : min=0 max=16 step=1 default=0 value=0
          display_pipeline_size 0x00981902 (int)    : min=0 max=6 step=1 default=3 value=3
       img_extra_decode_buffers 0x00981903 (int)    : min=-1 max=3 step=1 default=-1 value=-1
      override_spec_dpb_buffers 0x00981904 (int)    : min=-1 max=16 step=1 default=-1 value=-1

This allows complete control over the amount of memory allocated by the
driver.  Leaving the default values makes the driver allocate memory as
per the IMG hardware spec and the H.264 spec.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
(cherry picked from commit 9bfd5988373b8e13e8009bd3e453feb478b9c68f)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>