UPSTREAM: cpufreq: ti-cpufreq: Add opp support for am62p5 SoCs
The am62p5 is a variation of the am625 and the am62a7 SoC families. Add
the am62p5 to the list using the same cpufreq data as the rest of the
am62x extended family.

Change-Id: Id626a6cf6340bcfd301d62a1b0b6565badef85dc
Signed-off-by: Bryan Brattlof <bb@ti.com>
Reviewed-by: Dhruva Gole <d-gole@ti.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
(cherry picked from commit 304d9a0299f2afb54a699a9dcae7ed2f7fb20623)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>