TI: arm64: dts: ti: k3-am62p5-sk: Enable HDMI display on DSS0
The DSS0 outputs DPI signals via its second video port (VP2). The DPI
output from DSS0 is 24 bits (RGB888) and is forwarded to an HDMI
transmitter (SIL9022) on the board.

Add pinmux info for the DPI output.

Add DT nodes for SIL9022 HDMI transmitter (TX), and the HDMI connector
on the AM62P5 SK platform.

Additionally, connect the output of DSS0 (VP2) with input of the HDMI TX,
and the output of HDMI TX to the input of the HDMI connector.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
(cherry picked from commit 4cac4c7faeb762f6db55bd7d42fb916189193581)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>