TI: drm/tidss: Add dispc_is_idle()
Add a helper function, dispc_is_idle(), which returns whether the DSS is
idle (i.e. is any video port enabled).

For now we add a call to it in the suspend and resume callbacks, and
print a warning if in either place the DSS is not idle.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
(cherry picked from commit dc647bc9de78814b18533364ee9c35a2b74a5d62)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>