TI: counter/ti-eqep: add support for unit timer
This adds support to the TI eQEP counter driver for the Unit Timer.
The Unit Timer is a device-level extension that provides a timer to be
used for speed calculations. The sysfs interface for the Unit Timer is
new and will be documented in a later commit. It contains a R/W time
attribute for the current time, a R/W period attribute for the timeout
period and a R/W enable attribute to start/stop the timer. It also
implements a timeout event on the chrdev interface that is triggered
each time the period timeout is reached.

Link: https://www.spinics.net/lists/linux-iio/msg64817.html

Signed-off-by: David Lechner <david@lechnology.com>
[Judith: Fix merge conflicts. Only clear addressed interrupts. Add link to
commit message. Change clock to "fck". Fail probe if get clock rate fails]
Signed-off-by: Judith Mendez <jm@ti.com>
(cherry picked from commit d7c69d712ac8d62b59db1ec1f44168ea8801bce1)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>