TI: dt-bindings: usb/ti,am62-usb.yaml: Add vddcore supply
The USB driver needs to configure a bit in the USB PHY control
register depending on the PHY core voltage used in the system.

The PHY core supply is the same as the vdd_core supply on the SoC.

Add dt-bindings for this.

Signed-off-by: Roger Quadros <rogerq@kernel.org>
(cherry picked from commit c3aec8f5f52e583e018a9878cbd858939135736e)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>